requirements.txt
anthropic>=0.42.0
httpx>=0.28.0
mcp>=1.0.0
python-dotenv>=1.0.0
cml-mcp>=0.24.0
fastapi>=0.115.0
uvicorn>=0.34.0
ngrok>=1.0.0
